requirement.txt
# Core LangChain
langchain
langchain_community
langchainhub
langgraph
langgraph-checkpoint-sqlite

# LangChain Integrations
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
langchain_mistralai
langchain_huggingface

# AI/ML & Vector DB
chromadb
faiss-cpu
sentence_transformers
tiktoken

# Search & External APIs
wikipedia
google-search-results
tavily-python

# Google AI
google-generativeai

# Utilities
python-dotenv
spacy
textblob
emoji
streamlit